+++ src/settings/common.rs
@@ -1,7 +1,9 @@
+mod address;
 mod logging;
 mod network;
 mod value;
 
+pub use address::*;
 pub use logging::*;
 pub use network::*;
 pub use value::*;
